Add MARC record type
There's now a type in api.raml for a MARC record. It simply uses the
JSON-schema found at https://rossfsinger.com/blog/2010/09/
a-proposal-to-serialize-marc-in-json to serialize MARC in JSON.

It might be a good idea to one day translate the JSON-schema into a RAML
type, so that it wouldn't be so unreadable. It would be a VERY good idea
put all the RAML-referenced files in one directory. I'm-a gonna do that.