Tighten allowable keys in MARC control/subfields
The JSON schema from which the type was translated was a bit loose in
its allowable key names, and their numbers. As per the MARC-in-JSON
blog post (https://rossfsinger.com/blog/2010/09/
a-proposal-to-serialize-marc-in-json/), control/subfields must be in an
object containing exactly one property (the control/subfield), and of
course MARC imposes contraints on the size and makeup of control/
subfields. All this is now reflected in the RAML.

That reorganizing is coming soon, I promise!