Add Admin type in api.raml
Based loosely on admin.json, an admin has a set of permissions. Most of
these are simple read/write permissions for a resource type, but
checkouts are (kinda) special; they have another two permissions, signIn
and signOut, which are somewhat self-explanatory and are in addition to
checkout/(read|write).

Redundancy is reduced in the Admin type with a ReadWritePermission type,
used for the type of all read/write permissions.