Loosen permissions validation to appease raml2html
The (permissions) annotation now doesn't check whether the permission is
one of the allowable permissions; that information is now in the
description. The only purpose of permission-type.raml was to enable that
validation, so it was deleted.

This was necessary because raml2html (or raml2obj, or raml-js-parser)
has a bug where it doesn't substitute <<resourceTypeParameters>> before
validating annotations. Hopefully this works well enough and it doesn't
try to render the <<resourceTypeParameter>>, complete with double angle
brackets, as the annotation's value.