Add "unmodifiable" annotation
The annotation, in unmodifiable.raml, describes itself as indicating
"that this property is unmodifiable. Unmodifiable properties are
optional and must not be included in POST, PUT or PATCH request bodies,
but must be included in all responses bodies." It has no value.

This annotation was used in types.raml; hopefully it's not imported into
api.raml as types.unmodifiable...