Add .gitignore, remove api.html
Add a .gitignore. Currently it contains only api.html to ignore, but
once the implementation gets going it'll contain a lot more.

api.html was a generated file and should never have been in the repo
in the first place. For the record, it was generated by the npm module
raml2html@4.0.0-beta9 from raml/api.raml.