Add package.json for Node.js, update .gitignore
We're going to be using Node.js; added a package.json for a private
project with a bunch of useful dependencies. It's unlicensed for now,
this may or may not change later.

Also update the .gitignore to ignore some Node and NPM stuff. The new
contents of the .gitignore were generated by gitignore.io, a very useful
website.