Add Mongoose patron model
It's basically an exact clone of types.Patron from the API definition,
with the addition of an 'updated' Date property. Also, the URI is
validated for realsies this time, and not just with a regex.

Added valid-url as a dependency in package.json to use its URL
validation methods on pic's URI.