Add GET /v0/<model> in addCollection function
addCollection is a function in routes.js that will add the routes for
many endpoints for a model. It has 4 arguments (so far): the mongoose
model; the name of the collection, a string (e.g. "admins"); a function
that converts a document from the database into JSON to be sent to the
user, defaulting to the identity function (x => x); and whether the
collection is pageable, defaulting to true.

The GET method of /v0/<model> supports the sort_by and direction query
parameters for sorting (defaulting as specified in the RAML), and
optionally (but by default) pagination with the page and per_page query
parameters. It uses express-validator middleware for validation.

express-validator middleware was added in server.js, including a custom
sanitizer, setDefault. The sanitizer should make it easier to set
defaults for query parameters (hopefully it works).