Make unmodifiable check a middleware, reorganize
Added a middleware function for the unmodifiable property checking
that was previously done in the router.post body. This avoids code
duplication down the line. The middlware works only on POST, PUT, and
PATCH requests.

Also reorganized addCollection into middleware and routes. Previously,
middlware was interspersed throughout the function; now it's before
all the routes.