Add PUT /model/:id, make all model fields required
Add a PUT method to /v0/<model>/:id. It specially handles validation
errors and responds with a 422; this is to catch cases in which the
request didn't specify a field.

Make all model fields required in the model files. All fields are now
required; instead of, say:

    foo: String

there is now:

    foo: {type: String, required: true}

This is kinda ugly, but it should work for the PUT validation.