Pass next into hook to allow it to stop operation
Hooks now get an extra parameter, the `next` function from the
middleware. This allows the hook to decide whether to continue with the
operation, by calling next.

This is an improvement over the previous commit's method of giving the
hook this power, as you can't return from async code! Sigh, I want an
`await` keyword.