File-distribute routing logic
addCollection is now in routes/add-collection.js. It's wrapped in a
function that takes the router and the baseUri, which are the parameters
addCollection needs.

Utility functions common to both addCollection and the rest of the
routing logic are now in routes/util.js. This module is required by both
routes.js (which remains the main routing file), and add-collection.js.

This is for easier maintenance on the files.