Make handleDBError check whether there's an error
util.handleDBError no longer blithly assumes that when it's called,
its err param is truthy (i.e. there's an error). Instead, it now tests
whether there actually is an error before acting on that assumption. It
also returns true if there wasn't an error, and false otherwise (i.e. it
returns !err).

This allows us to turn code like this:

    someFunction('foo', function(err) {
      if (err) return util.handleDBError(err);
    });

into this:

    someFunction('foo', util.handleDBError);

which is much cleaner.