Add GET for /items/:id/marc, MARC/JSON converters
Add a GET path for /items/:id/marc, which works with both JSON and
application/marc. It simply finds the item in the database and returns
its marc property, possibly converting it to vanilla MARC.

Also add marc-convert.js, a helper module that exports two functions:
jsonToMarc and marcToJson. These functions do what they say on the tin:
they convert to and from MARC and JSON. It's required by routes.js for
the implementation of /items/:id/marc.