Add config, which is different in each environment
Add config.js, a file containing the configuration for each environment:
production, development, and test (for running unit tests). The config
changes depending on the value of NODE_ENV, an environment variable
containing the name of one of the three environments. Currently it is
only used for choosing the database -- library, library_dev, or
library_test.