Improve populateDB: callback args, genericization
populateDB now gives its success callback its `docs` argument (renamed
from `admins`). This is to reduce code duplication. It also gives the
callback the rest of the arguments it would normally get, but after
`docs`.

populateDB can also now accept a single doc with which to populate the
database, as well as an array. Reducing code duplication, again.

Finally, it is now less admin-centric, and can be exported in a util.js
file for tests. It does not use `admin` names, and instead uses `doc`.