Reduce code duplication in /v0/admins GET testing
Previously a lot of code was duplicated in the GET tests for /v0/admins.
Now it's not, as the duplicated code was extracted into a function named
testGet, and even the code calling *that* was extracted into
getGetTester. Now the body of the tests are just a simple call to
getGetTester. The testGet and getGetTester functions can be pulled out
into a 'util' testing file.