Add pageability tests
Pageability is now fully tested with ~20 tests, including error tests
(422, for invalid page/per_page, and 404, for paging beyond the end).
All tests pass, and all is right with the world.

We updated many of the testing methods, and added new ones. Added
testPaging and getPagingTester, which tests paging values returned.
getStatusTester now optionally takes documents to populate the DB with.
You can now generate admins with generateAdmins, a generator. Things now
sanitize the admins with a call to Array.from. Yeah.