Add PUT /v0/admins/:id tests
There's now a test helper function, testPut, and many tests using it
to test PUT /v0/admins/:id. All of them pass. In addition, the PUT
endpoint now tests whether the input is valid, then gives a 422 if it's
not (makes it idempotent).

testStatus now also checks for ':id' in the path, and if it's present
replaces it with the _id of the first doc passed.

Also added a dependency on chai-subset, which makes comparing the DB
and input objects a lot easier. It's used in testPut.