Add PATCH /v0/admins/:id error and success tests
Add a bunch of tests that test for invalid JSON, etc. handling. This
includes a bunch of tests that test what happens if someone tries to
delete any required property on the entire admin object, using a new
generator, generateSinglePropertyPatches, modeled on the similarly-long
named generateObjectsMissingOneProperty. It, as the name implies, simply
generates patches (with any value, defaulting to null) for each property
on an object (including nested properties) individually.

Also added a few success tests: the endpoint should ignore nonexistant
properties in the patches (it's not a bug it's a feature), and NOP on an
empty object.