Move all tests to template.js, call it in admin.js
All the tests have been moved to template.js, which exports a function
which takes an options parameter with a ton of arguments. When the
template function is called, it generates a metric tonne of tests
(~160 or so) for the endpoint specified. We did this in admin.js, which
is where everything was moved from.