Use template for patrons, add more hooks for it
Similarly to test/admin.js, added test/patron.js, which uses
test/template.js to generate ~150 tests. Simple enough, right? No.

Although it does not have as many properties, the patron resource is
more complex than the admin resource because patrons are interdependant
on other resources: item and checkout. This meant that custom hooks
had to be added to allow those checkouts and items to be added to the
database.

A summary of changes made:

- Two new options were added to template.js' function, both hooks:
  - customBeforeEach, a function called as a beforeEach hook, but with
    both `done` *and* `testDocs` as parameters. Why `testDocs`, you ask?
    It's because patron.js needed to modify the test document to add in
    the ObjectIds of the checkouts.
  - populateDBHooks, an object containing two hooks (`pre` and `post`)
    that are called pre and post populating the DB in populateDB().
- populateDB() was changed to take the hooks as parameters, which meant
  that everything that depends on populateDB() had to be changed to pass
  in the hooks, which meant that *they* needed to take the hooks as
  parameters, which meant that everything that called *them* needed to
  pass the hooks. Phew.
- Items were added in item-examples.js (the MARC records are so big
  that a whole file was needed). They're used in adding the items to the
  DB in patron.js.

So yeah, big commit. Gah.