Allow additional tests in the template
There is now one more template parameter: `additionalTests`. This
optional parameter is a function that will be called with no arguments
at the end of all the generated tests (but still in the describe()), if
it exists. This allows for custom tests to be inserted into the
template, but still under the describe().