Add MARC translations of items' JSON MARC records
test/test-docs/marc.js now contains a bunch of translations of the JSON
representations of the MARC contained in the item, now in proper MARC
format. test/test-docs/item.js was changed to remove all newlines from
the whitespace item (it wasn't playing nicely with the MARC format) and
fix a couple typos.