Add pagination support for patrons
You can now page between pages of patrons. Above the patron list is
information on the page number, maximum page, and patron numbers. Below
the patron list is links for next/previous/any absolute page number.

There's also a filter in core.js which, with the help of Stack Overflow,
will allow a range in the ng-repeat directive.