Add patron editing functionality
You can now edit a patron from the frontend by simply clicking on the
patron's hyperlinked name. It will open a form quite similar to the
"add a patron" form, but autofilled with the patron's information. The
client will then send a PUT request to /v0/patrons/:id with the new
patron, then reload the page of patrons.