Allow changing the number of patrons per page
There's now a series of links at the top of the page with every multiple
of ten from 10 to 100, which changes the number of patrons per page. The
client will try to keep the page the same, unless it'll run off the end
of the pages (in which case it'll go to the last page).