Allow hooks to pass new docs back to callers in add-collection.js
This is a bit complicated, m'kay? Previously, the hooks to
add-collection.js
could not change the docs at all. Now they can pass new docs into the
next()
function and they'll be used in place of the docs they were passed;
however,
if new docs are not passed into next(), the old docs will be used by
default.
This will be used for stuff like handling application/marc Content-Types
for
POST /v0/items.