Add a frontend for adding an item
There are two options for adding an item: you can fill out a form,
complete with a fancy MARC thing, or you can copy-paste a MARC record
in. Both work.

Note: we need to GENERATE A LEADER, currently every record entered via
the form has the leader '123456789012345678901234'. Clearly this is
wrong!!!